refactor(moonrun): expose embeddable runtime library
